requirements.txt
# libraries need to be installed 
streamlit
streamlit_extras
google_generativeai
python-dotenv
Pillow